+++ src/components/login/github-login.tsx
@@ -1,0 +1,11 @@
+const GitHubLoginBtn = () => {
+  return (
+    <button
+      className={`bg-black text-white text-lg text-center w-full focus:outline-none font-medium py-4 transition-colors`}
+    >
+      Login with Git Hub
+    </button>
+  );
+};
+
+export default GitHubLoginBtn;
