+++ src/styles/tailwind.css
@@ -1,20 +1,19 @@
 @tailwind base;
 @tailwind components;
 
-.input{
-    @apply focus:outline-none focus:border-gray-500 p-3 border-2 text-lg font-light border-gray-200
+.input {
+  @apply focus:outline-none focus:border-gray-500 p-3 border-2 text-lg font-light border-gray-200;
 }
 .container {
-    @apply max-w-screen-2xl mx-auto;
-  }
-  
-  .link {
-    @apply text-lime-600 hover:underline;
-  }
+  @apply max-w-screen-2xl mx-auto;
+}
 
-  .btn {
-    @apply text-lg font-medium focus:outline-none text-white py-4  transition-colors bg-lime-600 hover:bg-lime-700;
-  }
-  
+.link {
+  @apply text-lime-600 hover:underline;
+}
+
+.btn {
+  @apply text-lg font-medium focus:outline-none text-white py-4  transition-colors bg-lime-600 hover:bg-lime-700;
+}
 
 @tailwind utilities;
